Split client_models into multiple files in models directory
