+++ starknet_py/net/models/contracts.py
@@ -1,0 +1,59 @@
+from dataclasses import dataclass
+from typing import List, Dict, Any, Union, Iterable
+
+
+@dataclass
+class DeployedContract:
+    address: int
+    class_hash: int
+
+
+@dataclass
+class ContractCode:
+    """
+    Dataclass representing contract deployed to starknet
+    """
+
+    bytecode: List[int]
+    abi: List[Dict[str, Any]]
+
+
+@dataclass
+class EntryPoint:
+    """
+    Dataclass representing contract entry point
+    """
+
+    offset: int
+    selector: int
+
+
+@dataclass
+class EntryPointsByType:
+    """
+    Dataclass representing contract class entrypoints by entry point type
+    """
+
+    constructor: List[EntryPoint]
+    external: List[EntryPoint]
+    l1_handler: List[EntryPoint]
+
+
+@dataclass
+class DeclaredContract:
+    """
+    Dataclass representing contract declared to starknet
+    """
+
+    program: dict
+    entry_points_by_type: EntryPointsByType
+
+
+@dataclass
+class Call:
+    to_addr: int
+    selector: int
+    calldata: List[int]
+
+
+Calls = Union[Call, Iterable[Call]]
