+++ starknet_py/net/models/transaction_payloads.py
@@ -1,0 +1,166 @@
+from abc import ABC
+from dataclasses import dataclass, field
+from enum import Enum
+from typing import List, Optional
+
+from starknet_py.net.models.messages import L2toL1Message, L1toL2Message
+
+
+@dataclass
+class Transaction(ABC):
+    """
+    Dataclass representing common attributes of all transactions
+    """
+
+    hash: int
+    signature: List[int]
+    max_fee: int
+    version: int
+
+    def __post_init__(self):
+        if self.__class__ == Transaction:
+            raise TypeError("Cannot instantiate abstract Transaction class")
+
+
+@dataclass
+class InvokeTransaction(Transaction):
+    """
+    Dataclass representing invoke transaction
+    """
+
+    contract_address: int
+    calldata: List[int]
+    # This field is always None for transactions with version = 1
+    entry_point_selector: Optional[int] = None
+    nonce: Optional[int] = None
+
+
+@dataclass
+class DeclareTransaction(Transaction):
+    """
+    Dataclass representing declare transaction
+    """
+
+    class_hash: int
+    sender_address: int
+    nonce: Optional[int] = None
+
+
+@dataclass
+class DeployTransaction(Transaction):
+    """
+    Dataclass representing deploy transaction
+    """
+
+    contract_address: int
+    constructor_calldata: List[int]
+    class_hash: int
+
+
+@dataclass
+class DeployAccountTransaction(Transaction):
+    """
+    Dataclass representing deploy account transaction
+    """
+
+    contract_address_salt: int
+    class_hash: int
+    constructor_calldata: List[int]
+    nonce: int
+
+
+@dataclass
+class L1HandlerTransaction(Transaction):
+    """
+    Dataclass representing l1 handler transaction
+    """
+
+    contract_address: int
+    calldata: List[int]
+    entry_point_selector: int
+    nonce: Optional[int] = None
+
+
+class TransactionStatus(Enum):
+    """
+    Enum representing transaction statuses
+    """
+
+    NOT_RECEIVED = "NOT_RECEIVED"
+    RECEIVED = "RECEIVED"
+    PENDING = "PENDING"
+    ACCEPTED_ON_L2 = "ACCEPTED_ON_L2"
+    ACCEPTED_ON_L1 = "ACCEPTED_ON_L1"
+    REJECTED = "REJECTED"
+
+
+@dataclass
+class Event:
+    """
+    Dataclass representing an event emited by transaction
+    """
+
+    from_address: int
+    keys: List[int]
+    data: List[int]
+
+
+@dataclass
+class TransactionReceipt:
+    """
+    Dataclass representing details of sent transaction
+    """
+
+    # pylint: disable=too-many-instance-attributes
+
+    hash: int
+    status: TransactionStatus
+    block_number: Optional[int] = None
+    block_hash: Optional[int] = None
+    actual_fee: int = 0
+    rejection_reason: Optional[str] = None
+
+    events: List[Event] = field(default_factory=list)
+    l2_to_l1_messages: List[L2toL1Message] = field(default_factory=list)
+    l1_to_l2_consumed_message: Optional[L1toL2Message] = None
+
+
+@dataclass
+class SentTransactionResponse:
+    """
+    Dataclass representing a result of sending a transaction to starknet
+    """
+
+    transaction_hash: int
+    code: Optional[str] = None
+
+
+@dataclass
+class DeclareTransactionResponse(SentTransactionResponse):
+    """
+    Dataclass representing a result of declaring a contract on starknet
+    """
+
+    class_hash: int = 0
+
+
+@dataclass
+class DeployTransactionResponse(SentTransactionResponse):
+    """
+    Dataclass representing a result of deploying a contract to starknet
+    """
+
+    contract_address: int = 0
+
+
+@dataclass
+class TransactionStatusResponse:
+    block_hash: Optional[int]
+    transaction_status: TransactionStatus
+
+
+@dataclass
+class EstimatedFee:
+    overall_fee: int
+    gas_price: int
+    gas_usage: int
